# Reduce allocations when resolving object and disposing container