@ENikS 2018-11-20
Although proposed changes are beneficial to the performance I can not accept this PR. You should target vNext branch for all new development. I would be happy to merge your PR into vNext if you rewrite it.